backend/requirements.txt
amqp==2.1.4
billiard==3.5.0.2
celery==4.0.2
certifi==2017.4.17
chardet==3.0.4
coreapi==2.3.1
coreschema==0.0.4
defusedxml==0.5.0
Django==1.11.1
django-allauth==0.32.0
django-extensions==1.8.1
django-rest-auth==0.9.1
django-celery-results==1.0.1
django-rest-swagger==2.1.2
djangorestframework==3.6.3
djangorestframework-jwt==1.11.0
idna==2.5
itypes==1.1.0
Jinja2==2.9.6
kombu==4.0.2
MarkupSafe==1.0
oauthlib==2.0.2
openapi-codec==1.3.1
psycopg2==2.7.1
py==1.4.34
pytest==3.1.2
pytest-django==3.1.2
PyJWT==1.5.2
python3-openid==3.1.0
pytz==2017.2
redis==2.10.5
requests==2.18.1
requests-oauthlib==0.8.0
simplejson==3.11.1
six==1.10.0
uritemplate==3.0.0
urllib3==1.21.1
vine==1.1.3
